fix(qualita): evita segnalazioni vuote e corregge il conteggio nell'elenco recenti
- goBackToCommessa non resettava hasUnsaved: tornando indietro e poi avanti
  senza ridisegnare, "Termina segnalazione" salvava un disegno vuoto
  (trasparente) come segnalazione valida.
- "Salva segnalazione" ora richiede almeno un tratto disegnato.
- COUNT(*) in /reports/groups torna come stringa da Postgres: castato a
  ::int cosi' il singolare/plurale in italiano nell'elenco funziona.

+++ backend/src/modules/qualita/routes.ts
@@ -278,7 +278,7 @@ qualitaRoutes.get('/reports/groups', requireModule('qualita'), async (c) => {
     SELECT g.commessa, g.component_id, g.latest_at, g.count,
            cc.name AS component_name, cc.code AS component_code
     FROM (
-      SELECT commessa, component_id, MAX(created_at) AS latest_at, COUNT(*) AS count
+      SELECT commessa, component_id, MAX(created_at) AS latest_at, COUNT(*)::int AS count
       FROM qualita_report
       GROUP BY commessa, component_id
     ) g
